test: remove parameterized test helper
