+++ __tests__/__utils__/globalTestHelper.js
@@ -1,0 +1,13 @@
+'use strict'
+const {
+  getDefinition,
+  getLatestSupportedVersion,
+} = require('../../src/metadata/metadataManager')
+
+const getMetadata = async () => {
+  const apiVersion = await getLatestSupportedVersion()
+  const metadata = await getDefinition('directoryName', apiVersion)
+  return metadata
+}
+
+global.getGlobalMetadata = getMetadata
